+++ ext/Server/Bot/VehicleAttacking.lua
@@ -32,7 +32,7 @@ function VehicleAttacking:UpdateAttackingVehicle(p_Bot)
 						-- Chopper on Plane / Chopper --> weapon 2 (seaker)
 						if m_Vehicles:IsVehicleType(p_Bot.m_ActiveVehicle, VehicleTypes.Chopper) or
 							m_Vehicles:IsVehicleType(p_Bot.m_ActiveVehicle, VehicleTypes.Plane) then
-							if p_Bot.controlledEntryId == 0 and
+							if p_Bot.m_Player.controlledEntryId == 0 and
 								(p_Bot._ShootPlayerVehicleType == VehicleTypes.Chopper or p_Bot._ShootPlayerVehicleType == VehicleTypes.Plane) then
 								p_Bot._VehicleWeaponSlotToUse = 2
 							else