+++ ext/Shared/Constants/VehicleData.lua
@@ -27,7 +27,7 @@ VehicleData = {
 		Name = "[M1 ABRAMS]",
 		Type = VehicleTypes.Tank,
 		Terrain = VehicleTerrains.Land,
-		Parts = { { 25, 26 }, 12, nil }, -- 25,26 -- 12 -- none
+		Parts = { { 25, 26 }, 12, -1 }, -- 25,26 -- 12 -- none
 		Speed = { { 200, 610 }, 600, 350 },
 		Drop = { { 9.81, 9.81 }, 9.81, 9.81 }
 	},
@@ -35,7 +35,7 @@ VehicleData = {
 		Name = "[M1 ABRAMS]",
 		Type = VehicleTypes.Tank,
 		Terrain = VehicleTerrains.Land,
-		Parts = { { 25, 26 }, 12, nil }, -- 25,26 -- 12 -- none
+		Parts = { { 25, 26 }, 12, -1 }, -- 25,26 -- 12 -- none
 		Speed = { { 200, 610 }, 600, 350 },
 		Drop = { { 9.81, 9.81 }, 9.81, 9.81 }
 	},
@@ -43,7 +43,7 @@ VehicleData = {
 		Name = "[M1 ABRAMS]",
 		Type = VehicleTypes.Tank,
 		Terrain = VehicleTerrains.Land,
-		Parts = { { 25, 26 }, 12, nil }, -- 25,26 -- 12 -- none
+		Parts = { { 25, 26 }, 12, -1 }, -- 25,26 -- 12 -- none
 		Speed = { { 200, 610 }, 600, 350 },
 		Drop = { { 9.81, 9.81 }, 9.81, 9.81 }
 	},
@@ -51,7 +51,7 @@ VehicleData = {
 		Name = "[M1 ABRAMS]",
 		Type = VehicleTypes.Tank,
 		Terrain = VehicleTerrains.Land,
-		Parts = { { 25, 26 }, 12, nil }, -- 25,26 -- 12 -- none
+		Parts = { { 25, 26 }, 12, -1 }, -- 25,26 -- 12 -- none
 		Speed = { { 200, 610 }, 600, 350 },
 		Drop = { { 9.81, 9.81 }, 9.81, 9.81 }
 	},
@@ -59,7 +59,7 @@ VehicleData = {
 		Name = "[T-90A]",
 		Type = VehicleTypes.Tank,
 		Terrain = VehicleTerrains.Land,
-		Parts = { { 24, 25 }, 41, nil },
+		Parts = { { 24, 25 }, 41, -1 },
 		Speed = { { 200, 610 }, 600, 350 },
 		Drop = { { 9.81, 9.81 }, 9.81, 9.81 }
 	}, -- 24,25
@@ -67,7 +67,7 @@ VehicleData = {
 		Name = "[T-90A]",
 		Type = VehicleTypes.Tank,
 		Terrain = VehicleTerrains.Land,
-		Parts = { { 24, 25 }, 41, nil },
+		Parts = { { 24, 25 }, 41, -1 },
 		Speed = { { 200, 610 }, 600, 350 },
 		Drop = { { 9.81, 9.81 }, 9.81, 9.81 }
 	}, -- 24,25
@@ -75,7 +75,7 @@ VehicleData = {
 		Name = "[T-90A]",
 		Type = VehicleTypes.Tank,
 		Terrain = VehicleTerrains.Land,
-		Parts = { { 24, 25 }, 41, nil },
+		Parts = { { 24, 25 }, 41, -1 },
 		Speed = { { 200, 610 }, 600, 350 },
 		Drop = { { 9.81, 9.81 }, 9.81, 9.81 }
 	}, -- 24,25
@@ -83,7 +83,7 @@ VehicleData = {
 		Name = "[LAV-25]",
 		Type = VehicleTypes.Tank,
 		Terrain = VehicleTerrains.Amphibious,
-		Parts = { 10, 25, nil, nil, nil, nil },
+		Parts = { 10, 25, -1, -1, -1, -1 },
 		Speed = { 200, 610, 350, 350, 350, 350 },
 		Drop = { 4.9, 9.81, 9.81, 9.81, 9.81, 9.81 }
 	}, --10,19
@@ -91,7 +91,7 @@ VehicleData = {
 		Name = "[LAV-25]",
 		Type = VehicleTypes.Tank,
 		Terrain = VehicleTerrains.Land,
-		Parts = { 10, 25, nil, nil, nil, nil },
+		Parts = { 10, 25, -1, -1, -1, -1 },
 		Speed = { 200, 610, 350, 350, 350, 350 },
 		Drop = { 4.9, 9.81, 9.81, 9.81, 9.81, 9.81 }
 	}, --10,19
@@ -99,7 +99,7 @@ VehicleData = {
 		Name = "[LAV-25]",
 		Type = VehicleTypes.Tank,
 		Terrain = VehicleTerrains.Land,
-		Parts = { 10, 25, nil, nil, nil, nil },
+		Parts = { 10, 25, -1, -1, -1, -1 },
 		Speed = { 200, 610, 350, 350, 350, 350 },
 		Drop = { 4.9, 9.81, 9.81, 9.81, 9.81, 9.81 }
 	}, --10,19
@@ -107,7 +107,7 @@ VehicleData = {
 		Name = "[BTR-90]",
 		Type = VehicleTypes.Tank,
 		Terrain = VehicleTerrains.Land,
-		Parts = { 4, 29, nil, nil, nil, nil },
+		Parts = { 4, 29, -1, -1, -1, -1 },
 		Speed = { 200, 600, 350, 350, 350, 350 },
 		Drop = { 4.9, 9.81, 9.81, 9.81, 9.81, 9.81 }
 	},
@@ -115,7 +115,7 @@ VehicleData = {
 		Name = "[BMP-2M]",
 		Type = VehicleTypes.Tank,
 		Terrain = VehicleTerrains.Amphibious,
-		Parts = { 37, 49, nil, nil, nil, nil },
+		Parts = { 37, 49, -1, -1, -1, -1 },
 		Speed = { 200, 610, 350, 350, 350, 350 },
 		Drop = { 4.9, 9.81, 9.81, 9.81, 9.81, 9.81 }
 	}, --37,38
@@ -123,7 +123,7 @@ VehicleData = {
 		Name = "[BMP-2M]",
 		Type = VehicleTypes.Tank,
 		Terrain = VehicleTerrains.Amphibious,
-		Parts = { 37, 49, nil, nil, nil, nil },
+		Parts = { 37, 49, -1, -1, -1, -1 },
 		Speed = { 200, 610, 350, 350, 350, 350 },
 		Drop = { 4.9, 9.81, 9.81, 9.81, 9.81, 9.81 }
 	}, --37,38
@@ -133,7 +133,7 @@ VehicleData = {
 		Name = "[SPRUT-SD]",
 		Type = VehicleTypes.Tank,
 		Terrain = VehicleTerrains.Land,
-		Parts = { 16, nil, nil, nil, nil },
+		Parts = { 16, -1, -1, -1, -1 },
 		Speed = { 200, 350, 350, 350, 350 },
 		Drop = { 4.9, 9.81, 9.81, 9.81, 9.81 }
 	},
@@ -141,7 +141,7 @@ VehicleData = {
 		Name = "[M1128]",
 		Type = VehicleTypes.Tank,
 		Terrain = VehicleTerrains.Land,
-		Parts = { 26, nil, nil, nil, nil },
+		Parts = { 26, -1, -1, -1, -1 },
 		Speed = { 200, 350, 350, 350, 350 },
 		Drop = { 4.9, 9.81, 9.81, 9.81, 9.81 }
 	},
@@ -149,7 +149,7 @@ VehicleData = {
 		Name = "[RHINO]",
 		Type = VehicleTypes.Tank,
 		Terrain = VehicleTerrains.Land,
-		Parts = { 11, nil, nil, nil },
+		Parts = { 11, -1, -1, -1 },
 		Speed = { 600, 350, 350, 350 },
 		Drop = { 9.81, 9.81, 9.81, 9.81 }
 	},
@@ -185,7 +185,7 @@ VehicleData = {
 		Name = "[AAV-7A1 AMTRAC]",
 		Type = VehicleTypes.LightVehicle,
 		Terrain = VehicleTerrains.Amphibious,
-		Parts = { nil, 23, nil, nil, nil, nil },
+		Parts = { -1, 23, -1, -1, -1, -1 },
 		Speed = { 600, 80, 600, 600, 600, 600 },
 		Drop = { 9.81, 7.0, 9.81, 9.81, 9.81, 9.81 }
 	},
@@ -193,7 +193,7 @@ VehicleData = {
 		Name = "[M1114 HMMWV]",
 		Type = VehicleTypes.LightVehicle,
 		Terrain = VehicleTerrains.Land,
-		Parts = { nil, 19, nil, nil },
+		Parts = { -1, 19, -1, -1 },
 		Speed = { 300, 610, 300, 300 },
 		Drop = { 0.0, 9.81, 0.0, 0.0 }
 	},
@@ -201,7 +201,7 @@ VehicleData = {
 		Name = "[M1114 HMMWV]",
 		Type = VehicleTypes.LightVehicle,
 		Terrain = VehicleTerrains.Land,
-		Parts = { nil, 19, nil, nil },
+		Parts = { -1, 19, -1, -1 },
 		Speed = { 300, 610, 300, 300 },
 		Drop = { 0.0, 9.81, 0.0, 0.0 }
 	},
@@ -209,7 +209,7 @@ VehicleData = {
 		Name = "[M1114 HMMWV]",
 		Type = VehicleTypes.LightVehicle,
 		Terrain = VehicleTerrains.Land,
-		Parts = { nil, 19, nil, nil },
+		Parts = { -1, 19, -1, -1 },
 		Speed = { 300, 610, 300, 300 },
 		Drop = { 0.0, 9.81, 0.0, 0.0 }
 	},
@@ -217,7 +217,7 @@ VehicleData = {
 		Name = "[GAZ-3937 VODNIK]",
 		Type = VehicleTypes.LightVehicle,
 		Terrain = VehicleTerrains.Amphibious,
-		Parts = { nil, 23, nil, nil },
+		Parts = { -1, 23, -1, -1 },
 		Speed = { 300, 600, 300, 300 },
 		Drop = { 0.0, 9.81, 0.0, 0.0 }
 	},
@@ -225,7 +225,7 @@ VehicleData = {
 		Name = "[HMMWV ASRAD]",
 		Type = VehicleTypes.LightVehicle,
 		Terrain = VehicleTerrains.Land,
-		Parts = { nil, 25, nil, nil },
+		Parts = { -1, 25, -1, -1 },
 		Speed = { 300, 1000, 300, 300 },
 		Drop = { 300, 0.0, 300, 300 }
 	},
@@ -233,7 +233,7 @@ VehicleData = {
 		Name = "[VODNIK AA]",
 		Type = VehicleTypes.LightVehicle,
 		Terrain = VehicleTerrains.Land,
-		Parts = { nil, 12, nil, nil },
+		Parts = { -1, 12, -1, -1 },
 		Speed = { 300, 1000, 300, 300 },
 		Drop = { 9.81, 0.0, 9.82, 9.81 }
 	},
@@ -242,7 +242,7 @@ VehicleData = {
 		Name = "[BARSUK]",
 		Type = VehicleTypes.LightVehicle,
 		Terrain = VehicleTerrains.Land,
-		Parts = { nil, 6, 16 },
+		Parts = { -1, 6, 16 },
 		Speed = { 300, 600, 80 },
 		Drop = { 0.0, 15, 7.0 }
 	},
@@ -250,7 +250,7 @@ VehicleData = {
 		Name = "[PHOENIX]",
 		Type = VehicleTypes.LightVehicle,
 		Terrain = VehicleTerrains.Land,
-		Parts = { nil, 1, 18 },
+		Parts = { -1, 1, 18 },
 		Speed = { 300, 600, 80 },
 		Drop = { 0.0, 15, 7.0 }
 	},
@@ -260,7 +260,7 @@ VehicleData = {
 		Name = "[M142]", 
 		Type = VehicleTypes.MobileArtillery,
 		Terrain = VehicleTerrains.Land,
-		Parts = {nil, 13 },
+		Parts = {-1, 13 },
 		Speed = { 300, 50 },
 		Drop = { 0.0, 25.0}
 	},
@@ -268,7 +268,7 @@ VehicleData = {
 		Name = "[BM-23]",
 		Type = VehicleTypes.MobileArtillery,
 		Terrain = VehicleTerrains.Land,
-		Parts = {nil, 1 },
+		Parts = {-1, 1 },
 		Speed = { 300, 50 },
 		Drop = { 0.0, 25.0}
 	},
@@ -279,104 +279,104 @@ VehicleData = {
 		Name = "[A-10 THUNDERBOLT]",
 		Type = VehicleTypes.Plane,
 		Terrain = VehicleTerrains.Air,
-		Parts = { nil },
-		Speed = { 900 },
-		Drop = { 0.0 }
+		Parts = {  { -1, -1 } },
+		Speed = { {900, 10000} },
+		Drop = { {0.0, 0.0} }
 	},
 	["A10_THUNDERBOLT_spjet"] = {
 		Name = "[A-10 THUNDERBOLT]",
 		Type = VehicleTypes.Plane,
 		Terrain = VehicleTerrains.Air,
-		Parts = { nil },
-		Speed = { 900 },
-		Drop = { 0.0 }
+		Parts = {  { -1, -1 } },
+		Speed = { {900, 10000} },
+		Drop = { {0.0, 0.0} }
 	},
 	["F16"] = {
 		Name = "[F/A-18E SUPER HORNET]",
 		Type = VehicleTypes.Plane,
 		Terrain = VehicleTerrains.Air,
-		Parts = { nil },
-		Speed = { 900 },
-		Drop = { 0.0 }
+		Parts = {  { -1, -1 } },
+		Speed = { {900, 10000} },
+		Drop = { {0.0, 0.0} }
 	},
 	["F18_Wingman"] = {
 		Name = "[F/A-18E SUPER HORNET]",
 		Type = VehicleTypes.Plane,
 		Terrain = VehicleTerrains.Air,
-		Parts = { nil },
-		Speed = { 900 },
-		Drop = { 0.0 }
+		Parts = {  { -1, -1 } },
+		Speed = { {900, 10000} },
+		Drop = { {0.0, 0.0} }
 	},
 	["Su-25TM"] = {
 		Name = "[SU-25TM FROGFOOT]",
 		Type = VehicleTypes.Plane,
 		Terrain = VehicleTerrains.Air,
-		Parts = { nil },
-		Speed = { 900 },
-		Drop = { 0.0 }
+		Parts = {  { -1, -1 } },
+		Speed = { {900, 10000} },
+		Drop = { {0.0, 0.0} }
 	},
 	["Su-35BM Flanker-E"] = {
 		Name = "[SU-35BM FLANKER-E]",
 		Type = VehicleTypes.Plane,
 		Terrain = VehicleTerrains.Air,
-		Parts = { nil },
-		Speed = { 900 },
-		Drop = { 0.0 }
+		Parts = {  { -1, -1 } },
+		Speed = { {900, 10000} },
+		Drop = { {0.0, 0.0} }
 	},
 	["Su37"] = {
 		Name = "[SU-37]",
 		Type = VehicleTypes.Plane,
 		Terrain = VehicleTerrains.Air,
-		Parts = { nil },
-		Speed = { 900 },
-		Drop = { 0.0 }
+		Parts = {  { -1, -1 } },
+		Speed = { {900, 10000} },
+		Drop = { {0.0, 0.0} }
 	},
 	["F35B"] = {
 		Name = "[F-35]",
 		Type = VehicleTypes.Plane,
 		Terrain = VehicleTerrains.Air,
-		Parts = { nil },
-		Speed = { 900 },
-		Drop = { 0.0 }
+		Parts = {  { -1, -1 } },
+		Speed = { {900, 10000} },
+		Drop = { {0.0, 0.0} }
 	},
 	-- choppers
 	["AH1Z"] = {
 		Name = "[AH-1Z VIPER]",
 		Type = VehicleTypes.Chopper,
 		Terrain = VehicleTerrains.Air,
-		Parts = { nil, 1 }, --0,1,14
-		Speed = { 300, 600 },
-		Drop = { 0.0, 0.0 }
+		Parts = {  { -1, -1 }, 1 }, --0,1,14
+		Speed = { {300, 10000}, 600 },
+		Drop = { {0.0, 0.0}, 0.0 }
 	},
 	["AH1Z_coop"] = {
 		Name = "[AH-1Z VIPER]",
 		Type = VehicleTypes.Chopper,
 		Terrain = VehicleTerrains.Air,
-		Parts = { nil, 1 }, --0,1,14
-		Speed = { 300, 600 },
-		Drop = { 0.0, 0.0 }
+		Parts = {  { -1, -1 }, 1 }, --0,1,14
+		Speed = { {300, 10000}, 600 },
+		Drop = { {0.0, 0.0}, 0.0 }
 	},
 	["AH6_Littlebird"] = {
 		Name = "[AH-6J LITTLE BIRD]",
 		Type = VehicleTypes.Chopper,
 		Terrain = VehicleTerrains.Air,
-		Parts = { nil },
-		Speed = { 900 },
-		Drop = { 0.0 }
+		Parts = { { -1, -1 } },
+		Speed = { {900, 10000} },
+		Drop = { {0.0, 0.0} }
 	},
 	["AH6_Littlebird_EQ"] = {
 		Name = "[AH-6J LITTLE BIRD]",
 		Type = VehicleTypes.Chopper,
 		Terrain = VehicleTerrains.Air,
-		Parts = { nil, nil, nil, nil },
-		Speed = { 900, 300, 300, 300 },
-		Drop = { 0.0, 9.81, 9.81, 9.81 }
+		Parts = { { -1, -1 }, -1, -1, -1 },
+		Speed = { {900, 10000}, 300, 300, 300 },
+		Drop = { {0.0, 0.0}, 9.81, 9.81, 9.81 }
 	},
 	["Ka-60"] = {
 		Name = "[KA-60 KASATKA]",
 		Type = VehicleTypes.Chopper,
 		Terrain = VehicleTerrains.Air,
-		Parts = { nil, 18, 15, nil, nil },
+		Parts = { -1, 18, 15, -1, -1 },
 		Speed = { 350, 900, 900, 350, 350 },
 		Drop = { 9.81, 0.0, 0.0, 9.81, 9.81 }
 	},
@@ -384,15 +384,15 @@ VehicleData = {
 		Name = "[MI-28 HAVOC]",
 		Type = VehicleTypes.Chopper,
 		Terrain = VehicleTerrains.Air,
-		Parts = { nil, 2 }, --2,6},
-		Speed = { 350, 600 },
-		Drop = { 9.81, 0.0 }
+		Parts = {  { -1, -1 }, 2 }, --0,1,14
+		Speed = { {300, 10000}, 600 },
+		Drop = { {0.0, 0.0}, 0.0 }
 	},
 	["Venom"] = {
 		Name = "[UH-1Y VENOM]",
 		Type = VehicleTypes.Chopper,
 		Terrain = VehicleTerrains.Air,
-		Parts = { nil, 19, 16, nil, nil },
+		Parts = { -1, 19, 16, -1, -1 },
 		Speed = { 350, 900, 900, 350, 350 },
 		Drop = { 9.81, 0.0, 0.0, 9.81, 9.81 }
 	},
@@ -400,19 +400,33 @@ VehicleData = {
 		Name = "[UH-1Y VENOM]",
 		Type = VehicleTypes.Chopper,
 		Terrain = VehicleTerrains.Air,
-		Parts = { nil, 19, 16, nil, nil },
+		Parts = { -1, 19, 16, -1, -1 },
 		Speed = { 350, 900, 900, 350, 350 },
 		Drop = { 9.81, 0.0, 0.0, 9.81, 9.81 }
 	},
-	["Z-11w"] = { Name = "[Z-11W]", Type = VehicleTypes.Chopper, Terrain = VehicleTerrains.Air, Parts = {} },
-	["Wz11_SP_Paris"] = { Name = "[Z-11W]", Type = VehicleTypes.Chopper, Terrain = VehicleTerrains.Air, Parts = {} },
+	["Z-11w"] = {
+		Name = "[Z-11W]",
+		Type = VehicleTypes.Chopper, 
+		Terrain = VehicleTerrains.Air, 
+		Parts = {{-1, -1}, -1, -1}, -1,
+		Speed = { {900, 10000}, 350, 350, 350 },
+		Drop = { {0.0, 0.0}, 9.81, 9.81, 9.81 }
+	},
+	["Wz11_SP_Paris"] = { 
+		Name = "[Z-11W]",
+		Type = VehicleTypes.Chopper,
+		Terrain = VehicleTerrains.Air, 
+		Parts = {{-1, -1}, -1, -1}, -1,
+		Speed = { {900, 10000}, 350, 350, 350 },
+		Drop = { {0.0, 0.0}, 9.81, 9.81, 9.81 }
+	},
 
 	-- transport
 	["GrowlerITV"] = {
 		Name = "[GROWLER ITV]",
 		Type = VehicleTypes.NoArmorVehicle,
 		Terrain = VehicleTerrains.Land,
-		Parts = { nil, 47, nil },
+		Parts = { -1, 47, -1 },
 		Speed = { 300, 610, 300 },
 		Drop = { 0.0, 9.81, 0.0 }
 	},
@@ -420,7 +434,7 @@ VehicleData = {
 		Name = "[GROWLER ITV]",
 		Type = VehicleTypes.NoArmorVehicle,
 		Terrain = VehicleTerrains.Land,
-		Parts = { nil, 47, nil },
+		Parts = { -1, 47, -1 },
 		Speed = { 300, 610, 300 },
 		Drop = { 0.0, 9.81, 0.0 }
 	},
@@ -428,7 +442,7 @@ VehicleData = {
 		Name = "[VDV Buggy]",
 		Type = VehicleTypes.NoArmorVehicle,
 		Terrain = VehicleTerrains.Land,
-		Parts = { nil, 13, nil },
+		Parts = { -1, 13, -1 },
 		Speed = { 300, 610, 300 },
 		Drop = { 0.0, 9.81, 0.0 }
 	},
@@ -436,7 +450,7 @@ VehicleData = {
 		Name = "[DPV]",
 		Type = VehicleTypes.NoArmorVehicle,
 		Terrain = VehicleTerrains.Land,
-		Parts = { nil, 4, nil },
+		Parts = { -1, 4, -1 },
 		Speed = { 300, 610, 600 },
 		Drop = { 0.0, 9.81, 15 }
 	},
@@ -448,7 +462,7 @@ VehicleData = {
 	["Paris_SUV_Coop"] = { Name = "[SUV]", Type = VehicleTypes.NoArmorVehicle, Terrain = VehicleTerrains.Land, Parts = {} },
 	["Sniper_SUV"] = { Name = "[SUV]", Type = VehicleTypes.NoArmorVehicle, Terrain = VehicleTerrains.Land, Parts = {} },
 	["PoliceVan_Vehicle"] = { Name = "[POLICE VAN]", Type = VehicleTypes.NoArmorVehicle, Terrain = VehicleTerrains.Land, Parts = {} },
-	["RHIB"] = { Name = "[RHIB BOAT]", Type = VehicleTypes.NoArmorVehicle, Terrain = VehicleTerrains.Water,	Parts = { nil, nil, nil, nil } },
+	["RHIB"] = { Name = "[RHIB BOAT]", Type = VehicleTypes.NoArmorVehicle, Terrain = VehicleTerrains.Water,	Parts = { -1, -1, -1, -1 } },
 	["TechnicalTruck"] = { Name = "[TECHNICAL TRUCK]", Type = VehicleTypes.NoArmorVehicle, Terrain = VehicleTerrains.Land, Parts = {} },
 	["TechnicalTruck_Restricted"] = { Name = "[TECHNICAL TRUCK]", Type = VehicleTypes.NoArmorVehicle, Terrain = VehicleTerrains.Land, Parts = {} },